requirements.txt
Flask==3.1.0
pyrofork==2.3.58
python-dotenv==1.0.1
TgCrypto
requests
